backend/requirements.txt
openai==1.69.0
pinecone==5.4.2
langchain==0.3.21
langchain-community==0.3.20
langchain-openai==0.3.11
langchain-pinecone==0.2.3
sentence-transformers==4.0.1
fastapi
uvicorn
mangum
python-dotenv